Added unit test possibilities
