+++ .gitignore
@@ -1,4 +1,4 @@
-bin/*
-obj/*
+bin
+obj
 Parser.cs.old
 Scanner.cs.old
