+++ Matilda.Test/AbstractSyntax/Expr.Test.cs
@@ -1,0 +1,12 @@
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+
+namespace MatildaTests;
+
+[TestClass]
+public class ExprTest
+{
+    [TestMethod]
+    public void TestMethod1()
+    {
+    }
+}
