+++ Matilda.Test/AbstractSyntax/ExprTest.cs
@@ -1,0 +1,12 @@
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+
+namespace Company.TestProject1;
+
+[TestClass]
+public class Expr.Test
+{
+    [TestMethod]
+    public void TestMethod1()
+    {
+    }
+}
