+++ build.sh
@@ -1,2 +1,3 @@
+cd ./Matilda
 mono ./Coco.exe Matilda.cs.ATG -namespace Matilda
 dotnet build matilda.csproj -c Release
