--- matilda.sln
@@ -1,24 +1,0 @@
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.5.2.0
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "matilda", "matilda.csproj", "{0A999447-4C49-743B-96E6-EAC49CA1C060}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{0A999447-4C49-743B-96E6-EAC49CA1C060}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{0A999447-4C49-743B-96E6-EAC49CA1C060}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{0A999447-4C49-743B-96E6-EAC49CA1C060}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{0A999447-4C49-743B-96E6-EAC49CA1C060}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {ABFD107B-AEBC-4AA7-B2DE-70EDA6803602}
-	EndGlobalSection
-EndGlobal
