+++ p4-language.sln
@@ -1,0 +1,51 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.5.2.0
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Matilda", "Matilda\Matilda.csproj", "{09CEA7DD-D428-8B63-E051-3B224CB897C8}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MatildaTests", "Matilda.Test\MatildaTests.csproj", "{D72AF814-6DAB-41AC-A187-64B86BA0BB87}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|Any CPU = Release|Any CPU
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{09CEA7DD-D428-8B63-E051-3B224CB897C8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{09CEA7DD-D428-8B63-E051-3B224CB897C8}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{09CEA7DD-D428-8B63-E051-3B224CB897C8}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{09CEA7DD-D428-8B63-E051-3B224CB897C8}.Debug|x64.Build.0 = Debug|Any CPU
+		{09CEA7DD-D428-8B63-E051-3B224CB897C8}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{09CEA7DD-D428-8B63-E051-3B224CB897C8}.Debug|x86.Build.0 = Debug|Any CPU
+		{09CEA7DD-D428-8B63-E051-3B224CB897C8}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{09CEA7DD-D428-8B63-E051-3B224CB897C8}.Release|Any CPU.Build.0 = Release|Any CPU
+		{09CEA7DD-D428-8B63-E051-3B224CB897C8}.Release|x64.ActiveCfg = Release|Any CPU
+		{09CEA7DD-D428-8B63-E051-3B224CB897C8}.Release|x64.Build.0 = Release|Any CPU
+		{09CEA7DD-D428-8B63-E051-3B224CB897C8}.Release|x86.ActiveCfg = Release|Any CPU
+		{09CEA7DD-D428-8B63-E051-3B224CB897C8}.Release|x86.Build.0 = Release|Any CPU
+		{D72AF814-6DAB-41AC-A187-64B86BA0BB87}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D72AF814-6DAB-41AC-A187-64B86BA0BB87}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D72AF814-6DAB-41AC-A187-64B86BA0BB87}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{D72AF814-6DAB-41AC-A187-64B86BA0BB87}.Debug|x64.Build.0 = Debug|Any CPU
+		{D72AF814-6DAB-41AC-A187-64B86BA0BB87}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{D72AF814-6DAB-41AC-A187-64B86BA0BB87}.Debug|x86.Build.0 = Debug|Any CPU
+		{D72AF814-6DAB-41AC-A187-64B86BA0BB87}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D72AF814-6DAB-41AC-A187-64B86BA0BB87}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D72AF814-6DAB-41AC-A187-64B86BA0BB87}.Release|x64.ActiveCfg = Release|Any CPU
+		{D72AF814-6DAB-41AC-A187-64B86BA0BB87}.Release|x64.Build.0 = Release|Any CPU
+		{D72AF814-6DAB-41AC-A187-64B86BA0BB87}.Release|x86.ActiveCfg = Release|Any CPU
+		{D72AF814-6DAB-41AC-A187-64B86BA0BB87}.Release|x86.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {C73F0B88-2EA3-4B27-B3DD-538399CED17A}
+	EndGlobalSection
+EndGlobal
