+++ run.bat
@@ -1,2 +1,3 @@
 @echo off
+cd .\Matilda
 .\bin\Release\net10.0\matilda.exe code.matilda
