+++ run.sh
@@ -1,1 +1,2 @@
+cd ./Matilda
 dotnet ./bin/Release/net10.0/matilda.dll code.matilda
